+++ pipelex/pipe_controllers/pipe_batch.py
@@ -163,4 +163,5 @@ class PipeBatch(PipeController):
 
         return PipeOutput(
             working_memory=working_memory,
+            pipeline_run_id=job_metadata.pipeline_run_id,
         )