+++ tests/integration/pipelex/test_data.py
@@ -44,19 +44,16 @@ class PipeTestCases:
         name="text",
         concept_str="Text",
         content=TextContent(text="Describe a t-shirt in 2 sentences"),
-        pipelex_session_id="unit_test",
     )
     SIMPLE_STUFF_IMAGE = StuffFactory.make_stuff(
         name="image",
         concept_str="Image",
         content=ImageContent(url=URL_IMG_FASHION_PHOTO_1),
-        pipelex_session_id="unit_test",
     )
     SIMPLE_STUFF_PDF = StuffFactory.make_stuff(
         name="pdf",
         concept_str="PDF",
         content=PDFContent(url=PDFTestCases.DOCUMENT_URLS[0]),
-        pipelex_session_id="unit_test",
     )
     COMPLEX_STUFF = StuffFactory.make_stuff(
         name="complex",
@@ -67,7 +64,6 @@ class PipeTestCases:
                 ImageContent(url=URL_IMG_GANTT_1),
             ]
         ),
-        pipelex_session_id="unit_test",
     )
 
     STUFF_CONTENT_WITH_IMAGE_ATTRIBUTE_1 = SomeContentWithImageAttribute(image_attribute=ImageContent(url=URL_IMG_FASHION_PHOTO_1))
@@ -75,7 +71,6 @@ class PipeTestCases:
         concept_str="Image",
         content=STUFF_CONTENT_WITH_IMAGE_ATTRIBUTE_1,
         name="stuff_with_image",
-        pipelex_session_id="unit_test",
     )
     STUFF_CONTENT_WITH_IMAGE_ATTRIBUTE_IN_SUB_OBJECT = SomeContentWithImageSubObjectAttribute(
         image_attribute=ImageContent(url=URL_IMG_FASHION_PHOTO_2),
@@ -85,7 +80,6 @@ class PipeTestCases:
         concept_str="Image",
         content=STUFF_CONTENT_WITH_IMAGE_ATTRIBUTE_IN_SUB_OBJECT,
         name="stuff_with_image_in_sub_object",
-        pipelex_session_id="unit_test",
     )
     STUFFS_IMAGE_ATTRIBUTES: ClassVar[List[Tuple[Stuff, List[str]]]] = [  # stuff, attribute_paths
         (STUFF_WITH_IMAGE_ATTRIBUTE, ["stuff_with_image.image_attribute"]),
@@ -99,7 +93,7 @@ class PipeTestCases:
             ],
         ),
     ]
-    TRICKY_QUESTION_BLUEPRINT = StuffBlueprint(name="question", concept="answer.Question", value=USER_TEXT_TRICKY_2)
+    TRICKY_QUESTION_BLUEPRINT = StuffBlueprint(stuff_name="question", concept_code="answer.Question", content=USER_TEXT_TRICKY_2)
     BLUEPRINT_AND_PIPE: ClassVar[List[Tuple[str, StuffBlueprint, str]]] = [  # topic, blueprint, pipe
         (
             "Tricky question conclude",
@@ -148,7 +142,6 @@ class PipeTestCases:
                         TextContent(text="green"),
                     ]
                 ),
-                pipelex_session_id="unit_test",
             ),
             "colors",
             "color",